Add a note in the docs about thread safety of `#parse_wkt`
Since the parser is an ivar, and the parser has its own internal state, it's not safe to re-use an instance of a factory, e.g. as a class-level constant, to do parsing.

This is probably true for other methods as well, but `#parse_wkt` was the one I ran into this problem.