Geos ccw analysis (#229)
Add a `RGeo::Geos::Analysis.ccw?` method which itself uses the
GEOSCoordSeq_isCCW_r to determine if a geometry is counter-clockwise.
This method is used by `RGeo::Cartesian::Analysis.ccw?` when possible,
and there is a new `LinearRing#ccw?` method as well!

This geos version is only valid for the geos capi, and libgeos must be
3.7+.

Some refactoring was also done:

Add an error.h header to help throwing errors when there is an issue
in the c extension. And add a `RGeo::GeosError` for GEOS related
errors.

Change lots of `require` to `require_relative`. This helps a lot for
local development and clarifies usage.

Co-authored-by: Keith Doggett <keith.doggett887@gmail.com>