Add GeometryGraph class.
GeometryGraph is a subclass of PlanarGraph that is built from geometries instead of edges. GeometryGraph stores references to interior half-edges
for every component of the geometry. For example, a polygon will have one reference to a half-edge for it's exterior_ring and one for each
hole.

Also add, an #each method to HalfEdge that will follow next until next is nil or the calling half-edge.