Change validate_geometry to prepare_geometry and make it responsible for modifying geometry if necessary and no longer raising errors. Add validation test to all factories.
Before this commit, validate_geometry was responsible for 1) validating the geometries. 2) Modifying the geometries if necessary (ex. closing linear rings automatically).
Now it has been renamed to prepare_geometry and is only responsible for 2. Any validation logic that needs to be handled upon instantiation is handled in the constructor.

The CartesianValidOpTest was moved to a common ValidityTests module and all factories test against it. The spherical factory still needs some work since it doesn't have as
many geometric predicates implemented as the cartesian factory. There are also some issues where MultiPolygons are not properly instantiated in GEOS or GEOSFFI features.