Added Segment#segment_intersection to get points from intersections, created SweeplineIntersector class to compute intersections on groups of segments.
Segment#segment_intersection sligtly modifies the code from intersects_segment? and actually computes the intersection point (or nil if none).
Edge cases for colinear segments are handled by return a single point from the intersection.
intersects_segment? was modified to just check if segment_intersection is not nil.

A SweeplineIntersector class was added that uses a basic sweepline algorithm to compute the intersections in a set of segments.
This should be faster or as fast as the current validate_geometry intersection algorithm being used (more testing needed), since
it only compares segments where their y-range includes the y value of the event the sweepline is handling.
There are options to return all intersections or proper intersections that will filter out the connections in LineStrings.

In the future, it is possible to make a Sweepline class and have the SweeplineIntersector inherit from it. This would allow
us to use the sweepline more generically in other applications (some polygon clipping algos use a sweepline).