Add PlanarGraph Implementation
Adds a PlanarGraph implementation based on a Doubly Connected Edge List (DCEL). A DCEL is a graph data structure of
vertices and half-edges. Half-edges are directed edges where one segment can be represented by a pair of half-edges.
Half-edges are linked so that faces can be traversed (although faces are not stored).

The PlanarGraph accepts an array of segments to build it and additional edges can be added via the add_edge and add_edges methods.
When edges are added, intersections are automatically recomputed and handled. When possible, existing half-edges are preserved
so that geometries can reference those half-edges.

Also adds GeometryGraph which is a convenient way to interface geometries to a planar graph and provides ways to find important
half-edges in each geometry.

I haven't done a full performance test yet, but one notable area of improvement is that everytime an edge or edges is added,
all of the half-edges are re-linked when we likely only need to re-link the half-edges that originate at new vertices.