Add ImplHelper::ValidOp module to provide a uniform interface for validity checking of geometries.
The ValidOp module provides methods based on simple geometric predicates to determine validity for a geometry and can be used as a fallback or overridden.
The module can be overriden fully (in the case of Cartesian polygons) or just by implementing #invalid_reason (FFI and CAPI factories).

Removes #validate_geometry from all classes and replaces them with #prepare_geometry. Any required validations were moved to the constructors.

Adds the ValidityTest module to test/common to ensure that all classes of geometries validate in the same way.