Split ValidOp into ValidOp and ValidOpHelpers
ValidOp is to be included in feature classes and has the invalid_reason and valid? methods, as well as feature type specific methods that walk through the validity flow for each
geometry type.

ValidOpHelpers is a collection of functions where an object is passed to them. These perform the specific checks and return the invalid reason if applicable.

Also added overrides for both modules for cartesian functions.