Fix Psych::DisallowedClass errors in tests for ruby 3.1 / head
At the moment, tests for the latest ruby version are failing with a
Psych::DisallowedClass error. This is because Psych 4.0 aliases `load`
to `safe_load`.

(Related release notes: https://www.ruby-lang.org/en/news/2021/12/25/ruby-3-1-0-released/)

Using `unsafe_load` and ensuring that the latest version of
Psych is used, regardless of the ruby version, fixes this.