refactor: Remove `SRSDatabase` and `factory_generator` methods
There is no open source usage of SRSDatabase anywhere, see:

- https://grep.app/search?q=srs_database&filter[lang][0]=Ruby
- https://grep.app/search?q=SRSDatabase&filter[lang][0]=Ruby

As we'd like to simplify the codebase, we'll remove this for now, if
one wants it back, we're opened to extract this into a separate gem.

The only purpose of `factory_generator` methods was related to that, and
could simply be implemented using a `proc`, hence we also remove those
from the interface.