Remove Proj4 option from factory interface
In order to decouple the RGeo-Proj4 gem from the RGeo core gem, we first have to remove dependencies of factories on the proj gem for transforming geometries.
This commit replaces the function of the proj4 attribute in factories with the coord_sys attribute.
It also adds the feature that factories will try to create a coord_sys based on the given SRID if a coord_sys is not provided.
It adds the CoordSys::CONFIG.default_coord_sys_class that is used to try to instansiate coord_sys objects within a factory.
The idea is that the default_coord_sys_class will be reset once the proj4 gem is loaded.
Also adds the coord_sys_class option to factories so that a different adapter can be used if desired.