ci: add clang-format
Add a `bin/clang-format` command for practical usage, and this command
is also ran by the CI.

The `.clang-format` file is responsible of how the formatting is, it is
a bare copy of Mozilla's style guide for now, opiniated changes are
welcome if they provide clarifications.

This commit also applies clang-format.